% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bhlm.R
\name{bhlm.MAP}
\alias{bhlm.MAP}
\title{Maximum a posteriori}
\usage{
bhlm.MAP(bhlm_object, outcome_options = NULL,
  density.estimate = "logspline")
}
\arguments{
\item{bhlm_object}{Object returned from \code{bhlm}, of class \code{bhlm_object}.}

\item{outcome_options}{Character vector specifying which outcomes should be plotted.
Defaults to all outcome options from \code{bhlm_object@outcome_options}.}

\item{density.estimate}{Choose between \code{polspline::logspline} or \code{base::density} for density estimation method.}
}
\description{
Create a dataframe with outcome names, and maximum posteriori for each outcome.
You can choose the density estimation method (\code{polspline::logspline} or \code{base::density}).
}
\author{
Hugh Benjamin Zachariae
}
