% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bhlm.R
\name{bhlm}
\alias{bhlm}
\title{Bayesian Hierarchical Latent Mixture Model}
\usage{
bhlm(dataframe, grouping_factor_cols, estimate_col, outcome_options_col,
  outcome_options, outcome_priors, lambda_prior, theta_prior,
  field_theta_precision = NULL, identifier_col = NULL,
  bayes_method = "jags", jags_init = NULL, jags_chains = 3,
  jags_iter = 10000, jags_burnin = 1000, jags_thin = max(1,
  floor(jags_chains * (jags_iter - jags_burnin)/1000)), jags_DIC = TRUE,
  save_model = NULL)
}
\arguments{
\item{dataframe}{A data frame with studies, two grouping factor cols (e.g. studies*outcomes),
estimates, and optional ID column.}

\item{grouping_factor_cols}{A data vector of two grouping factor columns of the data frame
(e.g. \code{c("Study", "Outcome")}). These columns must be integer vectors.}

\item{estimate_col}{Column of estimates in the data frame.}

\item{outcome_options_col}{Column containing the different outcome names

 \code{Outcomes : Factor w/ 3 levels "Physical","Psychological",..: 2 2 2 2 2 2 2 2 1 1 ...}}

\item{outcome_options}{Chosen outcomes. You can choose a subset of outcomes.}

\item{outcome_priors}{Priors for outcomes. Takes character vector or \code{matrix}.

 Define all outcome priors to one distribution set prior (throws Warning) as \code{character}:

 \code{"dnorm(0, 1)"}

 \emph{Priors defined as character strings are checked with regex:}

 \code{grepl("^[a-zA-Z]+\\(((\\d*\\.)*\\d+,\\s?)*(\\d*\\.)*\\d+\\)", .)}

 or as data vector

   \code{c("dnorm", 0, 1)}

 Define all outcome priors manually with \code{matrix}. \strong{Notice}: byrow = TRUE

 \code{matrix(c("dnorm", 0, 1, "dnorm", 0, 1.1),
 nrow = 'number of outcomes',
 ncol = 'maximum amount of dist parameters',
 byrow = TRUE)}

 or as character vector of \code{characters}

 \code{c("dnorm(0,1)", "dnorm(0,1.1)")}

 \strong{Notice}: priors set manually, are set in the same order as \code{outcome_options}
 in character vector or in matrix row order (top to bottom).}

\item{lambda_prior}{Prior for Lambda as \code{character} or data vector.}

\item{theta_prior}{Prior for Theta as \code{character} or data vector.}

\item{identifier_col}{ID column. Chosen column is added to the used data for easier overview.}

\item{bayes_method}{Use JAGS or STAN (currently only JAGS is implemented).}

\item{save_model}{Input FALSE to not save the model file,
NULL to save as 'model_file.txt' in working directory, or
filepath location with name to save the model file as .txt (Separated with '\\' or '/').}

\item{field_theta}{Optional: Include a another hierarchical level in the form of a field-wide mean distribution.
Use a string vector with the same distribution as \code{theta_prior} and precision (no mean). The field-wide mean prior,
is set to \code{theta_prior}. This solution is currently a fix-up, and should only be used on a case by case basis.}

\item{jags}{Parameters for \code{R2jags::\link[R2jags]{jags}}:

  \code{init} (\code{NULL} set as default), \code{chains},
 \code{iter} (iterations per chain), \code{burning} (length of burn), \code{thin} (sim thinning rate), and
 \code{DIC} (compute deviance, pD and DIC).}
}
\description{
Run a BHLM model with custom number of latent variables according to specified number of outcome options.
The model sets priors for each outcome, group mean (theta) and estimate variance.

Run following for model specification (Graphical notation):

 \code{plot(0:1,0:1,type="n",ann=FALSE,axes=FALSE)}
 \code{rasterImage(bhlm_model_spec,0,0,1,1)}
}
\examples{

# INCOMING


}
\author{
Hugh Benjamin Zachariae
}
