% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bhlm_classes.R
\docType{class}
\name{bhlm_data-class}
\alias{bhlm_data-class}
\title{An S4 class to contain the used data for \code{latent_mixture} from the preprocessing.}
\description{
An S4 class to contain the used data for \code{latent_mixture} from the preprocessing.
}
\section{Slots}{

\describe{
\item{\code{used_data}}{The updated dataframe of the data, after preprocessing.}

\item{\code{start_bounds}}{Starting bounds between studies for the looping through estimates.}
}}

