% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{nativeRaster}
\name{bhlm_model_spec}
\alias{bhlm_model_spec}
\title{bhlm model specification jpeg}
\description{
Image displaying the graphical notation for the model specification.

Run following to show:

 \code{plot(0:1,0:1,type="n",ann=FALSE,axes=FALSE)}
 \code{rasterImage(bhlm_model_spec,0,0,1,1)}
}
\author{
Benjamin Zachariae, 2018.
}
\keyword{Bayesian,}
\keyword{Hierarchical}
\keyword{Latent}
\keyword{jpeg,}
\keyword{mixture}
\keyword{model,}
\keyword{model.}
