% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bhlm_classes.R
\docType{class}
\name{bhlm_object-class}
\alias{bhlm_object-class}
\title{An S4 class to contain the results and used data of \code{latent_mixture}.}
\description{
An S4 class to contain the results and used data of \code{latent_mixture}.
}
\section{Slots}{

\describe{
\item{\code{data}}{latent_mixture_data object, including the data used in the latent mixture model.}

\item{\code{jags_samples}}{rjags object from R2Jags.}
}}

