% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{CBT}
\alias{CBT}
\title{CBT data}
\description{
Data recieved with permission the authors.

Article:

Cognitive behavioral therapies for informal caregivers of
patients with cancer and cancer survivors: a systematic
review and meta-analysis.
}
\author{
Mia S. O'Toole, Robert Zachariae, Megan E. Renna,
Douglas S. Mennin, & Allison Applebaum, 2016
}
\keyword{behavoral}
\keyword{cognitive}
\keyword{meta-analysis,}
\keyword{therapy.}
