% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{EWI}
\alias{EWI}
\title{EWI data}
\description{
Data recieved with permission the authors.

Article:

The effect of expressive writing intervention on psychological
and physical health outcomes in cancer patients—a
systematic review and meta-analysis.
}
\author{
Robert Zachariae & Mia S. O'Toole, 2015
}
\keyword{expressive}
\keyword{meta-analysis,}
\keyword{writing}
