% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bhlm_plots.R
\name{bhlm.posteriorplot}
\alias{bhlm.posteriorplot}
\title{Show posterior distribution plots for outcomes}
\usage{
bhlm.posteriorplot(bhlm_object, return_plots = FALSE)
}
\arguments{
\item{bhlm_object}{Object returned from \code{bhlm}, of class \code{bhlm_object}.}

\item{return_plots}{Return ggplot objects in \code{list}.}
}
\description{
Plot outcomes with object from \code{bhlm}.
}
\author{
Hugh Benjamin Zachariae
}
